+++ js/main.js
@@ -1,0 +1,3 @@
+import {getPhotoAttributesArray} from './photo-attributes-generation.js';
+
+getPhotoAttributesArray();
