+++ js/photo-attributes-generation.js
@@ -1,14 +1,9 @@
+import getRandomInteger from './random-numbers-generation.js';
 
 const authorNames = ['Артем', 'Сергей', 'Дмитрий', 'Антон', 'Виктор'];
 const messages = ['Всё отлично!','В целом всё неплохо. Но не всё.','Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.','Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.','Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.','Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'];
 const descriptions = ['Ну, как-то так.', 'Не пытайтесь понять смысл этой фотограффии, я сам его не понимаю.', 'Это кошка.', 'Отпуск в самом разгаре.'];
 
-const getRandomInteger = (min, max) => {
-  const lower = Math.ceil(Math.min(min, max));
-  const upper = Math.floor(Math.max(min, max));
-  const result = Math.random() * (upper - lower + 1) + lower;
-  return Math.floor(result);
-};
 
 const getUnicRandomId = (min, max) => {
   const previousValues = [];
@@ -54,4 +49,4 @@ const getPhotoAttributesArray = () => {
   return photoAttributesArray;
 };
 
-getPhotoAttributesArray();
+export {getPhotoAttributesArray};
